rename supervisor module "citizen" to "supervision"
