+++ lib/modules/supervisor/supervision/is-name-registered-by-a-citizen.js
