+++ lib/modules/supervisor/supervision/notify-on-crash-or-shutdown.js
