+++ lib/modules/supervisor/supervision/notify-on-output.js
