+++ lib/modules/supervisor/supervision/register-citizen.js
