+++ lib/supervisor.js
@@ -10,7 +10,7 @@ function supe( config ){
   // load core modules
     supervisor.use( require( './modules/supervisor/noticeboard' ) );
     supervisor.use( require( './modules/supervisor/middleware' ) );
-    supervisor.use( require( './modules/supervisor/citizen' ) );
+    supervisor.use( require( './modules/supervisor/supervision' ) );
     supervisor.use( require( './modules/supervisor/signal' ) );
     supervisor.use( require( './modules/supervisor/mail' ) );
